Extract sync delta merge overlay
