+++ service-worker.js
@@ -146,6 +146,7 @@ const APP_SHELL = [
   '/js/sync-apply.js',
   '/js/sync-schema.js',
   '/js/sync-delta.js',
+  '/js/sync-delta-merge.js',
   '/js/sync-delta-registry.js',
   '/js/sync-delta-observability.js',
   '/js/sync-tombstones.js',
